special case is fixed
